Fix crash after closing of ejected device on Linux
When the user calls QSP::close() in a slot which is connected to the
QSP::error() signal, then the application is crashed. The reason is that
we emit the QSP::error() signal before than we call QRingBuffer::chop()
of an internal read buffer, which becomes invalid after the
QIODevice::close() called. Therefore, we need just call
QRingBuffer::chop() before than the QSP::error() signal will be emitted.

Task-number: QTBUG-55847
Change-Id: If536f9cf5cbc1b813d3642bdf9be0867e06368e8
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Alex Trotsenko <alex1973tr@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>