Don't poll for POLLIN event if device is open in WriteOnly on Linux
Following code snippet:

 QSerialPort serial("/dev/ttyUSB0");
 serial.open(QIODevice::WriteOnly);
 serial.write("ABCDEF");
 serial.waitForBytesWritten(-1);

causes an application crash if some of data will be received by the
serial port.

Reason is that qt_poll_msecs() triggered with POLLIN event, even if the
device is opened with O_WRONLY flag. In this case the readNotification()
handler is called, which trying to reserve some space from the
uninitialized read QRingBuffer, that causes an assertion.

Solution is to don't use the POLLIN event if device is open with
O_WRONLY flag.

Change-Id: I205be31608359f52b5ef286a33dd266ed11a2649
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>