Start overlapped I/O notifier before any asynchronous operation
This commit d57b9e83857253a1479f89fca18ea5856db39fb7 introduced
a regression, where the QSP may ignore all the read events when
the data comes to the device within opening. In this case, even
re-opening of a device does not help.

Reason is that the QWinOverlappedIoNotifier is enabled after than
the startAsyncCommunication() called, that probably, leads to
ignoring for all EV_RXCHAR events. A workaround is to enable the
notifier before than any of I/O operation called.

Task-number: QTBUG-67224
Change-Id: I21538fddc4dfd850c600e1d74f3c393a59a00549
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>