Fix shadow build qtserialbus
QtSerialBus requires QtSerialPort for its serial modbus feature. If
QtSerialPort is disabled the serial feature in QtSerialBus is disabled.
Integrity is one of the platforms where QtSerialPort is not supported
and therefore the serial modbus classes are removed.

The removal of QtSerialPort is done via qtserialport.pro. Unfortunately
during configure execution this project file is not considered. The
configure script only considers configure.json files to determine whether
a module is built/supported on a particular platform. Consequently
the build process fails to build QtSerialBus.

The fix is to copy the project conditions to QtSerialPort's configure.json.

Fixes: QTBUG-75279
Change-Id: I3796052de6c0622c334218e07ca4eab4134c9302
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>