QSerialPort: Port to alertable I/O functions on Windows
This commit removes the QWinOverlappedIoNotifier in favor of alertable
I/O functions like {Read|Write}FileEx(). The reason is that the
QWinOverlappedIoNotifier is very complex to maintain.

To use the alertable functions in the serial port we need an
alertable analog of WaitCommEvent function. This function does not
exist in the Win32 API, but we implement it trough a set of the
system NtXXX and RtlXXX functions, which are resolved dynamically.

This patch was tested with auto-tests and the examples using the:

* com0com virtual serial port driver.
* eltima virtual serial port driver.
* pl2303 USB/serial converters.

Task-number: QTBUG-74961
Change-Id: Idc428173eee7a1066a4693de00aa38416c4ee86c
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>