Avoid possible symbol clashes on static builds on Windows
We have same 'resolveSymbols' functions names in the global
namespace as in the qtserialport module and as in the qtserialbus
module. This can lead to the arbitrary call of this function from
a different module that can lead to the crash or other issue
(e.g. to this issue QTBUG-78546). So, we can just to rename this
function that solves this issue in the future.

Change-Id: I25f9d2a33c6858c5194d8c76eefd518eecfb6f7b
Reviewed-by: André Hartmann <aha_1980@gmx.de>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>