Update commercial license headers
Updated header.COMM to the files in tqtc-qtserialport.
Examples, tests or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
changes in the line count causes failures in tests.

Task-number: QTQAINFRA-4217
Change-Id: I35a7e264c70a923884ea0fec2c97ee824476062f
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>