Update commercial license headers
Updated header.COMM to the files in tqtc-qtserialport.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4961
Change-Id: I22444ff52c511a072dd507b0ccb948b995125524
Reviewed-by: Akseli Salovaara <akseli.salovaara@qt.io>