CMake: Make sure the Libudev package dependency is recorded
Otherwise only the target dependency is exported in
INTERFACE_LINK_LIBRARIES of SerialPort module, which means that
configuring qtserialbus in a static build could fail with one of
the following messages.

With CMake 3.21

 CMake Error at Qt6Core/Qt6CoreMacros.cmake:2468 (add_library):
  Target "SerialBus" links to target "PkgConfig::Libudev" but the
  target was not found.  Perhaps a find_package() call is missing
  for an IMPORTED target, or an ALIAS target is missing?

With CMake 3.23

CMake Error at Qt6SerialPort/Qt6SerialPortTargets.cmake:51
  (set_target_properties):
  The link interface of target "Qt6::SerialPort" contains:

    PkgConfig::Libudev

  but the target was not found.  Possible reasons include:

    * There is a typo in the target name.
    * A find_package call is missing for an IMPORTED target.
    * An ALIAS target is missing.

By specifying PROVIDED_TARGETS, we ensure that
Qt6SerialPortDependencies.cmake will call find_dependency(Libudev).

Note this only fails in the tqtc repo, and not the regular repo, due
to qt_internal_read_repo_dependencies populating tqtc repo names
from dependencies.yaml in QT_REPO_DEPENDENCIES, which is used in
QtPluginConfig.cmake.in.

In the non-tqtc repo, the dependencies.yaml value of 'qtbase' matches
the recorded lower case 'QtBase' project when loading the
DeviceDiscovery static plugin, which has a dependency on Libudev.
So the Libudev package was implicitly / accidentally found.

In the tqtc repo, dependencies.yaml has 'tqtc-qtbase', which does not
match 'QtBase', thus a lot of static plugin Config files are not loaded,
including the ones that have an implicit dependency on Libudev,
resulting in the issue of SerialBus project not configuring.

This is a quick correct fix to resolve this particular problem.
But we also need a fix in QtPluginConfig.cmake.in which will consider
normalized repo names, like it was done in qt5 in
30068f2223b112650a1fb61ca6fd3b42cb26fcd3

Task-number: QTBUG-102883
Change-Id: I037b8f873f92dc0b0709430140d6fea86e032fb8
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>