Windows: fix soft memory leak in synchronous mode
QWinOverlappedIoNotifier is using an atomic 'waiting' variable to make
sure that it does not emit _q_notify() signals in synchronous mode.
However this does not work in practice, because a notification from the
I/O completion port thread can come before we enter the blocking waiting
function, and the signal will be emitted anyway.
Those signals migth never processed when using synchronous read/write,
because in such scenarios QSerialPort works in an infinite loop in its
own thread, probably without leaving a change for proper event
processing. This leads to a situation when a lot of signals are queued
on the event loop, and the memory consumption grows.

This patch introduces another atomic variable to make sure that a new
signal is never emitted until the previous one is processed. With such
approach we have maximum one signal pending on the queue.
The drawback of this approach is that when working in asynchronous
mode, notifications might come faster than we process them, so the
processing queue might grow. To solve this, we update the code that is
handles notifications in asynchronous mode to process the whole input
queue instead of only one item.

This commit was manually picked from
01e0371f9d69943e62619fdbfe6a9d87f2a2fa16

Fixes: QTBUG-101444
Fixes: QTBUG-103822
Fixes: QTBUG-93865
Fixes: QTBUG-91237
Fixes: QTBUG-87151
Change-Id: Iad3c14ea5f0d3f3f3f9d483a1e6ab3e8b3c6c573
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>
(cherry picked from commit f3a306a30fc4f40d1c96fee0ed44517fe8b43d76)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>