Fix bindable properties
It looks like the bindable property implementation has changed after
the support for bindable properties was added to the module.
And because of the lack of testing in the CI, we failed to notice that
the tests started to fail.

This patch fixes the bindable properties implementation.

Fixes: QTBUG-109219
Change-Id: Id2677d6dfb74f26881c4ae73a844136fff160b02
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 71791866d3e167b89d8fc94162fa778e6b577ddb)