Doc: Use the \examplecategory macro to tag examples
The '\meta category' command was used for tagging examples with a
specific category, used in Qt Creators Welcome mode.

As we want to also generate lists of examples belonging to a category
elsewhere in the documentation, replace the command with a macro that
expands to the original \meta command and also adds the example to
a group using the \ingroup command. This way, the category names can
be used as arguments to the \generatelist or \annotatedlist commands.

Task-number: QTBUG-112731
Change-Id: I207110715b0d3bab25d21ef91f7b0c833c0acb27
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit cfeca8077184f175fb5278d682b89ef1ad83bbfd)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>