SerialPort: fix bindable properties
Update the bindable property tests to use the helper functions from
QTestPrivate. The only tricky case here is the breakEnabled property
which requires the port to be opened. So, craft a lambda that creates
a helper serial port, opens it, and sets a proper default value.

This revealed that all bindable properties had the same issue - the
setter was using property.value(), thus creating a binding loop.
Fix it by using valueBypassingBindings() instead.

Task-number: QTBUG-116544
Pick-to: 6.6 6.5
Change-Id: I28de7598207737ea97a2ec8fd94dc65603421e5c
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>