QSerialPortInfo::isNull(): fix the usage of inline keyword
The inline keyword should be used on the function declaration, not on
the definition [0].
Alternatively, the function can be implemented directly in-class, in
which case the inline keyword should be omitted.

Since the isNull() function is a one-liner, use the second approach.

[0]: https://lists.qt-project.org/pipermail/development/2023-August/044351.html

Fixes: QTBUG-120221
Pick-to: 6.5 5.15
Change-Id: I39697ccec81f3da7d0b303c31f4d1b0c03018f1a
Reviewed-by: Tatiana Borisova <tatiana.borisova@qt.io>
(cherry picked from commit fca9174d4935756d8c0b669b36f62d47eb1b071a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>