Doc: clarify QSerialPort::open() error reporting
Unlike most other Qt classes, QSerialPort emits an errorOccurred()
signal with NoError error code if an attempt to open the port is
successful. We even have unit-tests that rely on this behavior.

This is somewhat confusing, but this behavior has been around at least
since Qt 5.9 (see the linked Jira ticket). As this is mostly harmless,
and some people can actually rely on this behavior, do not change it,
but rather extend the docs to mention this corner case.

As a drive-by, drop the \note from another paragraph in this function's
docs to avoid having too many notes in the same place.

Fixes: QTBUG-67544
Pick-to: 6.8 6.5 5.15
Change-Id: Ied4e02e4ab293f04f5527c8df1a2550264b843cb
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit b444675656055b4019d775f80fff4a7aa28111dd)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>