Add .gitreview file
`git-review` is a command line tool for interacting with Gerrit.
Some IDEs offer support for this tool, either natively or through
plugins. The tool relies on a configuration file, .gitreview. Adding
this configuration file to our repositories simplifies initial setup
for contributors that use the tool directly or through their IDE of
choice.

The configuration file adds a remote called 'gerrit'. This is the
default for the tool, and also the name set for
codereview.qt-project.org by Qt's `init-repository` script. Thus,
the configuration should work seamlessly alongside other repository
helpers.

As the file format doesn't seem to be designed with comments in mind,
add it to the other version control system exemptions in REUSE.toml.

Task-number: QTBUG-132604
Pick-to: 6.9 6.8
Change-Id: I0369a8f109b9a71050f8f4a7d3725a409749a00d
Reviewed-by: Samuel Gaist <samuel.gaist@idiap.ch>